refactor: extract instance types and state constants
